backend/requirements.txt
# Core API
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0

# Neo4j Knowledge Graph
neo4j==5.25.0
py2neo==2021.2.4

# LLM & Embeddings
openai==1.50.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0
langgraph==0.2.28

# Milvus (Dhruv - RAG vector store)
pymilvus==2.4.7

# NLP Pipeline (CrowdSignal)
spacy==3.7.6
beautifulsoup4==4.12.3
requests==2.32.3

# Crowd Score Model
lightgbm==4.5.0
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4

# Scheduling & Notifications
apscheduler==3.10.4
firebase-admin==6.5.0

# Weather
# (uses requests - already installed)

# Whisper (voice - add later if time permits)
# openai-whisper==20231117
